requirements.txt
langchain
pypdf2 
python-dotenv
streamlit
langchain-openai
tiktoken
faiss-cpu